XDG Work
Added support for XDG_CONFIG_HOME and XDG_CACHE_HOME.  If not set,
defaults to ~/.config/styli.sh and ~/.cache/styli.sh and will create the
dirs if they do not exist.

Moved the download to $XDG_CACHE_HOME/styli.sh and in all cases called
'wallpaper.jpg'.

The subreddit list needs to be present in $XDG_CONFIG_HOME/styli.sh and
will let the user know if it cannot find it.